Live Demo

+++ phi/tf/model.py
@@ -43,7 +43,7 @@ class FieldSequenceModel(nontf.FieldSequenceModel):
     def editable_int(self, name, initial_value, minmax=None):
         val = EditableInt(name, initial_value, minmax, None)
         setattr(self, 'int_'+name.lower(), val)
-        placeholder = tf.placeholder(tf.float32, (), name.lower().replace(' ', '_'))
+        placeholder = tf.placeholder(tf.int32, (), name.lower().replace(' ', '_'))
         self.add_scalar(name, placeholder)
         self.editable_placeholders[placeholder] = 'int_'+name.lower()
         return placeholder
